修复动态、直播不能解析用户id为long时的bug

+++ BilibiliMonitor/Models/DynamicModel.cs
@@ -64,7 +64,7 @@
             public object following { get; set; }
             public string jump_url { get; set; }
             public string label { get; set; }
-            public int mid { get; set; }
+            public long mid { get; set; }
             public string name { get; set; }
             public Official_Verify official_verify { get; set; }
             public Pendant pendant { get; set; }
